refactoring

+++ src/main/java/forex/processing/NumberFormatter.java
@@ -7,11 +7,11 @@ public class NumberFormatter {
     private NumberFormatter() {
     }
 
-    public static float round(float value) {
+    public static float round(double value) {
         return NumberFormatter.round(value, 5);
     }
 
-    public static float round(float value, int numberOfDigitsAfterDecimalPoint) {
+    public static float round(double value, int numberOfDigitsAfterDecimalPoint) {
         BigDecimal bigDecimal = new BigDecimal(value);
         bigDecimal = bigDecimal.setScale(numberOfDigitsAfterDecimalPoint, BigDecimal.ROUND_HALF_UP);
         return bigDecimal.floatValue();
